Edited LPIs
Images of license plates are edited accordingly to task. Task is to process images of license plates of parked cars. Images that are for processing now are from cars in motion or on random roads and streets. Need to find more images of parked cars with a clear view of license plates.