Recognition Prep
Preparing for recognition of license plate in Recognition.py. Function should segment off regions from thresholded image, and crop the image to according regions. These cropped images should be added to a list/map, and the flattened for KNN recognition. All of this data should be added to a map, for ease of use when recognizing / mapping / drawing on original / segmeting original.