Filtering
Filtering results of cropped image regions. Filters for now include
(1) - Counting how many times a letter has been found. If this number is 1, the letter will not even be considered valid.
(2) - Checking the KNN distance. If this distance is over 6kk it will not be considered valid
(3) - Counting these valid distances. If this count is under 1, it will not be considered. (NOTE: Seems redundant, but this way we skip printing and checking letters counted over 1 time.)